Add universal trait for GetByBioguideId Queries
