+++ src/php/audit.congress/objects/members/querys/class.member.offices.query.php
@@ -3,19 +3,14 @@
 namespace AuditCongress {
 
     class MemberOfficesQuery extends AuditCongressQuery {
+        use GetByBioguideIdQuery;
+
         public function __construct() {
             parent::__construct("MemberOffices");
         }
 
         public function applyDefaultOrder() {
             $this->setOrderBy(["officeId"], false);
-        }
-
-        public static function getByBioguideId($bioguideId) {
-            $offices = new MemberOfficesQuery();
-            $offices->setSearchColumns(["bioguideId"]);
-            $offices->setSearchValues([$bioguideId]);
-            return $offices->selectFromDB()->fetchAllAssoc();
         }
 
         public static function getByOfficeId($officeId) {
