+++ src/php/audit.congress/objects/members/querys/class.member.socials.query.php
@@ -3,19 +3,14 @@
 namespace AuditCongress {
     
     class MemberSocialsQuery extends AuditCongressQuery {
+        use GetByBioguideIdQuery;
+
         public function __construct() {
             parent::__construct("MemberSocials");
         }
 
         public function applyDefaultOrder() {
             $this->setOrderBy(["bioguideId"], false);
-        }
-
-        public static function getByBioguideId($bioguideId) {
-            $socials = new MemberSocialsQuery();
-            $socials->setSearchColumns(["bioguideId"]);
-            $socials->setSearchValues([$bioguideId]);
-            return $socials->selectFromDB()->fetchAllAssoc();
         }
     }
 }
