+++ src/php/audit.congress/objects/members/querys/class.member.terms.query.php
@@ -3,19 +3,14 @@
 namespace AuditCongress {
 
     class MemberTermsQuery extends AuditCongressQuery {
+        use GetByBioguideIdQuery;
+        
         public function __construct() {
             parent::__construct("MemberTerms");
         }
 
         public function applyDefaultOrder() {
             $this->setOrderBy(["start"], false);
-        }
-
-        public static function getByBioguideId($bioguideId) {
-            $terms = new MemberTermsQuery();
-            $terms->setSearchColumns(["bioguideId"]);
-            $terms->setSearchValues([$bioguideId]);
-            return $terms->selectFromDB()->fetchAllAssoc();
         }
 
         private static function getSingleTermByBioguideId($bioguideId) {
